Update versions of specimen and eid. Update script to modify specimen and eid
