+++ bia-ingest/bia_ingest/conversion/experimental_imaging_dataset.py
@@ -113,7 +113,7 @@ def get_experimental_imaging_dataset(
             "analysis_method": analysis_method_list,
             "correlation_method": correlation_method_list,
             "example_image_uri": [],
-            "version": 1,
+            "version": 0,
             "attribute": {"associations": associations},
         }
         model_dict["uuid"] = generate_experimental_imaging_dataset_uuid(model_dict)
